requirements.txt
aiogram==2.25
aiohttp==3.8.6
annotated-types==0.6.0
async-timeout==4.0.3
charset-normalizer==3.3.2
coloredlogs==15.0.1
executor==23.2
frozenlist==1.4.1
fsspec==2024.2.0
grpcio==1.62.0
grpcio-status==1.62.0
huggingface-hub==0.21.1
humanfriendly==10.0
idna==3.6
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
packaging==23.2
pillow==10.2.0
property-manager==3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
tortoise-orm==0.20.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
verboselogs==1.7
yarl==1.9.4
